feat: Refactor Canvas component into modular hooks
- Remove monolithic Canvas.tsx and Node.tsx components
- Extract canvas event handlers into useCanvasEventHandlers hook
- Extract port-related handlers into usePortHandlers hook
- Extract keyboard shortcut handlers into useKeyboardHandlers hook
- Simplify Canvas/index.tsx by integrating new hooks and smaller components
- Improve code maintainability and readability by modularizing logic
- Preserve existing functionality including node/edge management, drag/drop,
  predicted nodes, schema synchronization, and keyboard shortcuts
