+++ src/app/canvas/components/Canvas/hooks/index.ts
@@ -4,6 +4,9 @@ export { useNodeManagement } from './useNodeManagement';
 export { useEdgeManagement } from './useEdgeManagement';
 export { useDragState } from './useDragState';
 export { usePredictedNodes } from './usePredictedNodes';
+export { useCanvasEventHandlers } from './useCanvasEventHandlers';
+export { usePortHandlers } from './usePortHandlers';
+export { useKeyboardHandlers } from './useKeyboardHandlers';
 
 export type { 
     UseCanvasViewReturn,
