+++ src/app/canvas/components/Canvas/index.tsx
@@ -20,25 +20,17 @@ import {
     useNodeManagement,
     useEdgeManagement,
     useDragState,
-    usePredictedNodes
+    usePredictedNodes,
+    useCanvasEventHandlers,
+    usePortHandlers,
+    useKeyboardHandlers
 } from './hooks';
 
 // Components
 import { CanvasNodes, CanvasPredictedNodes, CanvasEdges } from './components';
 
 // Utils
-import {
-    areTypesCompatible,
-    validateRequiredInputs,
-    generatePortKey,
-    parsePortKey,
-    findPortData,
-    isParameterInput,
-    getWorldPosition,
-    isClick,
-    findClosestSnapTarget,
-    SNAP_DISTANCE
-} from './utils';
+import { areTypesCompatible, validateRequiredInputs } from './utils';
 
 // Types
 import type {
@@ -202,7 +194,7 @@ const Canvas = forwardRef<CanvasRef, CanvasProps>(({
 
     // Port reference registration
     const registerPortRef = useCallback((nodeId: string, portId: string, portType: string, el: HTMLElement | null) => {
-        const key = generatePortKey(nodeId, portId, portType as 'input' | 'output');
+        const key = `${nodeId}__PORTKEYDELIM__${portId}__PORTKEYDELIM__${portType}`;
         
         if (el) {
             portRefs.current.set(key, el);
@@ -284,8 +276,9 @@ const Canvas = forwardRef<CanvasRef, CanvasProps>(({
         addEdge, 
         isDraggingOutput, 
         isDraggingInput, 
-        sourcePortForConnection, 
-        areTypesCompatible
+        sourcePortForConnection,
+        setEdgePreview,
+        setSourcePortForConnection
     ]);
 
     // Schema synchronization
@@ -359,476 +352,82 @@ const Canvas = forwardRef<CanvasRef, CanvasProps>(({
         devLog.log('Schema synchronization completed successfully');
     }, [nodes, edges, setNodes]);
 
-    // Event Handlers
-    const handleCanvasMouseDown = (e: React.MouseEvent<HTMLDivElement>): void => {
-        const target = e.target as HTMLElement;
-        if (isParameterInput(target)) return;
-        if (e.button !== 0) return;
-
-        if (isDraggingOutput || isDraggingInput) {
-            clearPredictedNodes();
-        }
-
-        clearSelection();
-        startCanvasDrag(e, view);
-    };
-
-    const handleMouseMove = useCallback((e: React.MouseEvent<HTMLDivElement>): void => {
-        if (dragState.type === 'none') return;
-
-        // Clear port click start when dragging begins
-        if (portClickStart) {
-            setPortClickStart(null);
-        }
-
-        if (dragState.type === 'canvas') {
-            setView(prev => ({ 
-                ...prev, 
-                x: e.clientX - (dragState.startX || 0), 
-                y: e.clientY - (dragState.startY || 0) 
-            }));
-        } else if (dragState.type === 'node') {
-            const newX = (e.clientX / view.scale) - (dragState.offsetX || 0);
-            const newY = (e.clientY / view.scale) - (dragState.offsetY || 0);
-            
-            setNodes(prevNodes =>
-                prevNodes.map(node =>
-                    node.id === dragState.nodeId ? { ...node, position: { x: newX, y: newY } } : node
-                )
-            );
-        } else if (dragState.type === 'edge') {
-            const container = containerRef.current;
-            if (!container || !edgePreviewRef.current) return;
-
-            const rect = container.getBoundingClientRect();
-            const mousePos = getWorldPosition(e.clientX, e.clientY, rect, view);
-
-            setEdgePreview(prev => prev ? { ...prev, targetPos: mousePos } : null);
-
-            // Find closest snap target
-            const edgeSource = edgePreviewRef.current.source;
-            if (!edgeSource) return;
-
-            const closestPortKey = findClosestSnapTarget(
-                mousePos,
-                portPositions,
-                SNAP_DISTANCE,
-                (key) => {
-                    const parsed = parsePortKey(key);
-                    if (!parsed) return false;
-                    
-                    // Filter logic for valid snap targets
-                    return parsed.portType === 'input' && edgeSource.nodeId !== parsed.nodeId;
-                }
-            );
-
-            if (closestPortKey) {
-                const parsed = parsePortKey(closestPortKey);
-                if (parsed) {
-                    let targetPort = null;
-                    
-                    if (isNodePredicted(parsed.nodeId)) {
-                        const predictedNode = predictedNodes.find(pNode => pNode.id === parsed.nodeId);
-                        if (predictedNode?.nodeData.inputs) {
-                            targetPort = predictedNode.nodeData.inputs.find(port => port.id === parsed.portId);
-                        }
-                    } else {
-                        targetPort = findPortData(nodes, parsed.nodeId, parsed.portId, parsed.portType);
-                    }
-                    
-                    const isValid = targetPort ? areTypesCompatible(edgeSource.type, targetPort.type) : false;
-                    setIsSnapTargetValid(isValid);
-                }
-            } else {
-                setIsSnapTargetValid(true);
-            }
-
-            setSnappedPortKey(closestPortKey);
-        }
-    }, [
-        dragState, 
-        view, 
-        portPositions, 
-        predictedNodes, 
-        nodes, 
-        portClickStart, 
-        setView, 
-        setNodes, 
-        setEdgePreview, 
-        setSnappedPortKey, 
-        setIsSnapTargetValid, 
-        setPortClickStart
-    ]);
-
-    const handleMouseUp = useCallback((e?: React.MouseEvent<HTMLDivElement>): void => {
-        if (dragState.type === 'edge' && (isDraggingOutput || isDraggingInput) && !snappedPortKeyRef.current && e) {
-            const container = containerRef.current;
-            if (container && edgePreviewRef.current) {
-                const rect = container.getBoundingClientRect();
-                const mousePos = getWorldPosition(e.clientX, e.clientY, rect, view);
-
-                const sourceType = edgePreviewRef.current.source.type;
-                let predicted: any[] = [];
-
-                if (isDraggingOutput) {
-                    predicted = generatePredictedNodes(sourceType, mousePos);
-                } else if (isDraggingInput) {
-                    predicted = generatePredictedOutputNodes(sourceType, mousePos);
-                }
-
-                setPredictedNodes(predicted);
-                setEdgePreview(null);
-                stopDrag();
-                return;
-            }
-        }
-
-        stopDrag();
-
-        if (dragState.type === 'edge' && snappedPortKeyRef.current) {
-            const source = edgePreviewRef.current?.source;
-            const parsed = parsePortKey(snappedPortKeyRef.current);
-            
-            if (source && parsed) {
-                handlePortMouseUp({
-                    nodeId: parsed.nodeId,
-                    portId: parsed.portId,
-                    portType: parsed.portType as 'input' | 'output',
-                    type: '' // Will be filled by handlePortMouseUp
-                });
-            }
-        }
-
-        setEdgePreview(null);
-        setSnappedPortKey(null);
-        setIsSnapTargetValid(true);
-    }, [
-        dragState, 
-        isDraggingOutput, 
-        isDraggingInput, 
-        view, 
-        generatePredictedNodes, 
-        generatePredictedOutputNodes, 
-        setPredictedNodes, 
-        setEdgePreview, 
-        stopDrag, 
-        setSnappedPortKey, 
-        setIsSnapTargetValid
-    ]);
-
-    const handleNodeMouseDown = useCallback((e: React.MouseEvent, nodeId: string): void => {
-        if (e.button !== 0) return;
-        selectNode(nodeId);
-        const node = nodes.find(n => n.id === nodeId);
-        if (node) {
-            startNodeDrag(e, nodeId, node.position, view);
-        }
-    }, [nodes, view, selectNode, startNodeDrag]);
-
-    const handleEdgeClick = useCallback((edgeId: string): void => {
-        selectEdge(edgeId);
-    }, [selectEdge]);
-
-    const handlePortMouseDown = useCallback((data: PortMouseEventData, mouseEvent?: React.MouseEvent): void => {
-        const { nodeId, portId, portType, isMulti, type } = data;
-
-        if (mouseEvent) {
-            setPortClickStart({
-                data,
-                timestamp: Date.now(),
-                position: { x: mouseEvent.clientX, y: mouseEvent.clientY }
-            });
-        }
-
-        if (portType === 'input') {
-            let existingEdge: CanvasEdge | undefined;
-            if (!isMulti) {
-                existingEdge = edges.find(e => e.target.nodeId === nodeId && e.target.portId === portId);
-            } else {
-                existingEdge = edges.findLast(e => e.target.nodeId === nodeId && e.target.portId === portId);
-            }
-            
-            if (existingEdge) {
-                startEdgeDrag();
-                const sourcePosKey = generatePortKey(existingEdge.source.nodeId, existingEdge.source.portId, existingEdge.source.portType as 'input' | 'output');
-                const sourcePos = portPositions[sourcePosKey];
-                const targetPosKey = generatePortKey(existingEdge.target.nodeId, existingEdge.target.portId, existingEdge.target.portType as 'input' | 'output');
-                const targetPos = portPositions[targetPosKey];
-
-                const sourcePortData = findPortData(nodes, existingEdge.source.nodeId, existingEdge.source.portId, existingEdge.source.portType);
-
-                if (sourcePos && sourcePortData) {
-                    setEdgePreview({
-                        source: { ...existingEdge.source, type: sourcePortData.type },
-                        startPos: sourcePos,
-                        targetPos: targetPos
-                    });
-                }
-
-                removeEdge(existingEdge.id);
-                return;
-            } else {
-                startEdgeDrag();
-                setIsDraggingInput(true);
-                setCurrentInputType(type);
-                setSourcePortForConnection({ nodeId, portId, portType, type });
-
-                const startPosKey = generatePortKey(nodeId, portId, portType as 'input' | 'output');
-                const startPos = portPositions[startPosKey];
-                if (startPos) {
-                    setEdgePreview({
-                        source: { nodeId, portId, portType, type },
-                        startPos,
-                        targetPos: startPos
-                    });
-                }
-                return;
-            }
-        }
-
-        if (portType === 'output') {
-            startEdgeDrag();
-            setIsDraggingOutput(true);
-            setCurrentOutputType(type);
-            setSourcePortForConnection({ nodeId, portId, portType, type });
-
-            const startPosKey = generatePortKey(nodeId, portId, portType as 'input' | 'output');
-            const startPos = portPositions[startPosKey];
-            if (startPos) {
-                setEdgePreview({ 
-                    source: { nodeId, portId, portType, type }, 
-                    startPos, 
-                    targetPos: startPos 
-                });
-            }
-            return;
-        }
-    }, [
-        edges, 
-        portPositions, 
-        nodes, 
-        startEdgeDrag, 
-        setPortClickStart, 
-        setEdgePreview, 
-        removeEdge, 
-        setIsDraggingInput, 
-        setIsDraggingOutput, 
-        setCurrentInputType, 
-        setCurrentOutputType, 
-        setSourcePortForConnection
-    ]);
-
-    const handlePortMouseUp = useCallback((data: PortMouseEventData, mouseEvent?: React.MouseEvent): void => {
-        const { nodeId, portId, portType, type } = data;
-        const currentEdgePreview = edgePreviewRef.current;
-
-        // Handle click vs drag detection
-        const isClickAction = portClickStart && mouseEvent &&
-            isClick(
-                portClickStart.timestamp,
-                portClickStart.position,
-                { x: mouseEvent.clientX, y: mouseEvent.clientY }
-            );
-
-        // Handle port click (generate predicted nodes)
-        if (isClickAction && 
-            portClickStart.data.nodeId === nodeId &&
-            portClickStart.data.portId === portId &&
-            portClickStart.data.portType === portType) {
-            
-            const portPosKey = generatePortKey(nodeId, portId, portType as 'input' | 'output');
-            const portPos = portPositions[portPosKey];
-
-            if (portPos) {
-                setSourcePortForConnection({ nodeId, portId, portType, type });
-                setEdgePreview({
-                    source: { nodeId, portId, portType, type },
-                    startPos: portPos,
-                    targetPos: portPos
-                });
-
-                let predicted: any[] = [];
-                if (portType === 'output') {
-                    setIsDraggingOutput(true);
-                    setCurrentOutputType(type);
-                    predicted = generatePredictedNodes(type, portPos);
-                } else if (portType === 'input') {
-                    setIsDraggingInput(true);
-                    setCurrentInputType(type);
-                    predicted = generatePredictedOutputNodes(type, portPos);
-                }
-
-                if (predicted.length > 0) {
-                    setPredictedNodes(predicted);
-                }
-            }
-
-            setPortClickStart(null);
-            return;
-        }
-
-        setPortClickStart(null);
-
-        if (!currentEdgePreview) return;
-
-        // Validate connection
-        if (!areTypesCompatible(currentEdgePreview.source.type, type)) {
-            setSnappedPortKey(null);
-            setIsSnapTargetValid(true);
-            setEdgePreview(null);
-            return;
-        }
-
-        if (currentEdgePreview.source.portType === portType) {
-            setEdgePreview(null);
-            return;
-        }
-
-        if (currentEdgePreview.source.nodeId === nodeId) {
-            setEdgePreview(null);
-            return;
-        }
-
-        // Handle predicted node connection
-        if (isNodePredicted(nodeId)) {
-            // Convert predicted node and connect
-            const predictedNode = predictedNodes.find(pNode => pNode.id === nodeId);
-            if (predictedNode) {
-                const newNode: CanvasNode = {
-                    id: `${predictedNode.nodeData.id}-${Date.now()}`,
-                    data: predictedNode.nodeData,
-                    position: predictedNode.position
-                };
-
-                addNode(newNode);
-
-                let newEdge: CanvasEdge;
-                if (isDraggingOutput) {
-                    newEdge = {
-                        id: `edge-${currentEdgePreview.source.nodeId}:${currentEdgePreview.source.portId}-${newNode.id}:${portId}-${Date.now()}`,
-                        source: currentEdgePreview.source,
-                        target: { nodeId: newNode.id, portId, portType }
-                    };
-                } else {
-                    newEdge = {
-                        id: `edge-${newNode.id}:${portId}-${currentEdgePreview.source.nodeId}:${currentEdgePreview.source.portId}-${Date.now()}`,
-                        source: { nodeId: newNode.id, portId, portType: 'output' },
-                        target: currentEdgePreview.source
-                    };
-                }
-
-                addEdge(newEdge);
-                clearPredictedNodes();
-            }
-            
-            setEdgePreview(null);
-            setSnappedPortKey(null);
-            setIsSnapTargetValid(true);
-            return;
-        }
-
-        // Handle regular node connection
-        clearPredictedNodes();
-
-        const edgeSignature = `${currentEdgePreview.source.nodeId}:${currentEdgePreview.source.portId}-${nodeId}:${portId}`;
-        if (isDuplicateEdge(currentEdgePreview.source.nodeId, currentEdgePreview.source.portId, nodeId, portId)) {
-            setEdgePreview(null);
-            return;
-        }
-
-        let newEdges = [...edges];
-        if (portType === 'input') {
-            const targetPort = findPortData(nodes, nodeId, portId, 'input');
-            if (targetPort && !targetPort.multi) {
-                newEdges = newEdges.filter(edge => !(edge.target.nodeId === nodeId && edge.target.portId === portId));
-            }
-        }
-
-        let newEdge: CanvasEdge;
-        if (currentEdgePreview.source.portType === 'input') {
-            newEdge = {
-                id: `edge-${nodeId}:${portId}-${currentEdgePreview.source.nodeId}:${currentEdgePreview.source.portId}-${Date.now()}`,
-                source: { nodeId, portId, portType },
-                target: currentEdgePreview.source,
-            };
-        } else {
-            newEdge = {
-                id: `edge-${edgeSignature}-${Date.now()}`,
-                source: currentEdgePreview.source,
-                target: { nodeId, portId, portType }
-            };
-        }
-
-        setEdges([...newEdges, newEdge]);
-        setEdgePreview(null);
-        setSnappedPortKey(null);
-        setIsSnapTargetValid(true);
-    }, [
-        portClickStart,
+    // Event Handlers using custom hooks
+    const canvasHandlers = useCanvasEventHandlers({
+        dragState,
+        view,
         portPositions,
-        edges,
         nodes,
         predictedNodes,
         isDraggingOutput,
         isDraggingInput,
-        areTypesCompatible,
-        isNodePredicted,
-        isDuplicateEdge,
-        findPortData,
-        addNode,
-        addEdge,
-        clearPredictedNodes,
-        setPortClickStart,
-        setSourcePortForConnection,
+        portClickStart,
+        containerRef,
+        edgePreviewRef,
+        snappedPortKeyRef,
+        setView,
+        setNodes,
         setEdgePreview,
+        setSnappedPortKey,
+        setIsSnapTargetValid,
+        setPortClickStart,
+        setPredictedNodes,
+        clearSelection,
+        startCanvasDrag,
+        clearPredictedNodes,
+        generatePredictedNodes,
+        generatePredictedOutputNodes,
+        stopDrag,
+        selectNode,
+        startNodeDrag,
+        selectEdge,
+        areTypesCompatible,
+        findPortData: (nodes: CanvasNode[], nodeId: string, portId: string, portType: string) => {
+            const node = nodes.find(n => n.id === nodeId);
+            if (!node) return null;
+            const portList = portType === 'input' ? node.data.inputs : node.data.outputs;
+            return portList?.find(p => p.id === portId) || null;
+        },
+        isNodePredicted,
+        handlePortMouseUp: () => {} // Will be provided by portHandlers
+    });
+
+    const portHandlers = usePortHandlers({
+        edges,
+        nodes,
+        predictedNodes,
+        portPositions,
+        isDraggingOutput,
+        isDraggingInput,
+        portClickStart,
+        edgePreviewRef,
+        setPortClickStart,
+        setEdgePreview,
+        setSnappedPortKey,
+        setIsSnapTargetValid,
         setIsDraggingOutput,
         setIsDraggingInput,
         setCurrentOutputType,
         setCurrentInputType,
+        setSourcePortForConnection,
         setPredictedNodes,
+        setEdges,
+        startEdgeDrag,
+        removeEdge,
+        addNode,
+        addEdge,
+        clearPredictedNodes,
+        isDuplicateEdge,
         generatePredictedNodes,
         generatePredictedOutputNodes,
-        setSnappedPortKey,
-        setIsSnapTargetValid,
-        setEdges
-    ]);
+        isNodePredicted
+    });
 
-    // Keyboard shortcuts
-    const handleKeyDown = useCallback((e: KeyboardEvent): void => {
-        const target = e.target as HTMLElement;
-        if (target.tagName === 'INPUT' || target.tagName === 'SELECT' || target.tagName === 'TEXTAREA') {
-            return;
-        }
-
-        const isCtrlOrCmd = e.ctrlKey || e.metaKey;
-
-        if (isCtrlOrCmd && e.key === 'c') {
-            e.preventDefault();
-            if (selectedNodeId) copyNode(selectedNodeId);
-        } else if (isCtrlOrCmd && e.key === 'v') {
-            e.preventDefault();
-            const pastedNodeId = pasteNode();
-            if (pastedNodeId) selectNode(pastedNodeId);
-        } else if (isCtrlOrCmd && e.key === 'z') {
-            e.preventDefault();
-            const restoredNode = undoDelete();
-            if (restoredNode) {
-                const connectedEdges = lastDeleted?.edges || [];
-                setEdges(prev => [...prev, ...connectedEdges]);
-            }
-        } else if ((e.key === 'Delete' || e.key === 'Backspace') && selectedNodeId) {
-            e.preventDefault();
-            const connectedEdges = removeNodeEdges(selectedNodeId);
-            deleteNode(selectedNodeId, connectedEdges);
-            clearSelection();
-        } else if ((e.key === 'Delete' || e.key === 'Backspace') && selectedEdgeId) {
-            e.preventDefault();
-            removeEdge(selectedEdgeId);
-            clearSelection();
-        }
-    }, [
+    const keyboardHandlers = useKeyboardHandlers({
         selectedNodeId,
         selectedEdgeId,
+        lastDeleted,
+        setEdges,
         copyNode,
         pasteNode,
         undoDelete,
@@ -836,10 +435,24 @@ const Canvas = forwardRef<CanvasRef, CanvasProps>(({
         removeEdge,
         removeNodeEdges,
         clearSelection,
-        selectNode,
-        lastDeleted,
-        setEdges
-    ]);
+        selectNode
+    });
+
+    // Get final handlers from hooks
+    const { 
+        handleCanvasMouseDown,
+        handleMouseMove,
+        handleMouseUp,
+        handleNodeMouseDown,
+        handleEdgeClick
+    } = canvasHandlers;
+
+    const { 
+        handlePortMouseDown,
+        handlePortMouseUp
+    } = portHandlers;
+
+    const { handleKeyDown } = keyboardHandlers;
 
     // Imperative handle for parent component access
     useImperativeHandle(ref, () => ({
